youtube-summary-transcription/requirements.txt
python-dotenv
langgraph
langchain-mcp-adapters
langchain_google_genai
asyncio
google.genai
fastmcp